refactor: argparser types refers to options types
